+++ components/ProductCardEarn.tsx
@@ -7,8 +7,7 @@ import { Flex, Text } from 'theme-ui'
 import { formatCryptoBalance, formatPercent } from '../helpers/formatters/format'
 import { ProductCardData, productCardsConfig } from '../helpers/productCards'
 import { one } from '../helpers/zero'
-import { calculateTokenAmount, ProductCard } from './ProductCard'
-import { ProductCardProtocolLink } from './ProductCardProtocolLink'
+import { calculateTokenAmount, ProductCard, ProductCardProtocolLink } from './ProductCard'
 import { StatefulTooltip } from './Tooltip'
 
 interface UnprofitableTooltipProps {
