+++ components/ProductCardMultiply.tsx
@@ -5,8 +5,7 @@ import React from 'react'
 import { formatCryptoBalance, formatPercent } from '../helpers/formatters/format'
 import { ProductCardData, productCardsConfig } from '../helpers/productCards'
 import { one } from '../helpers/zero'
-import { calculateTokenAmount, ProductCard } from './ProductCard'
-import { ProductCardProtocolLink } from './ProductCardProtocolLink'
+import { calculateTokenAmount, ProductCard, ProductCardProtocolLink } from './ProductCard'
 
 function personaliseCardData({
   productCardData,
